Rename CreateStore.js to createStore.js
