+++ app/redux/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux'
-import configureStore from 'app/redux/CreateStore'
+import configureStore from 'app/redux/createStore'
 
 export default () => {
   const rootReducer = combineReducers({})
